+++ HamiltonIO/epw/__init__.py
@@ -1,2 +1,10 @@
-from .epwparser import Epmat, read_crystal_fmt
-from .plot import plot_epw_distance
+from .epwparser import (
+    Epmat as Epmat,
+)
+from .epwparser import (
+    read_crystal_fmt as read_crystal_fmt,
+)
+from .epwparser import (
+    validate_epw_ws_weights as validate_epw_ws_weights,
+)
+from .plot import plot_epw_distance as plot_epw_distance